Fix test timing
