+++ packages-exp/analytics-exp/src/factory.ts
@@ -20,9 +20,6 @@ import {
   SettingsOptions,
   DynamicConfig,
   MinimalDynamicConfig,
-  CustomParams,
-  AnalyticsCallOptions,
-  EventParams,
   Analytics
 } from '@firebase/analytics-types-exp';
 import {
@@ -31,13 +28,6 @@ import {
   wrapOrCreateGtag,
   findGtagScriptOnPage
 } from './helpers';
-import {
-  logEvent as internalLogEvent,
-  setCurrentScreen as internalSetCurrentScreen,
-  setUserId as internalSetUserId,
-  setUserProperties as internalSetUserProperties,
-  setAnalyticsCollectionEnabled as internalSetAnalyticsCollectionEnabled
-} from './functions';
 import { AnalyticsError, ERROR_FACTORY } from './errors';
 import { FirebaseInstallations } from '@firebase/installations-types-exp';
 import { areCookiesEnabled, isBrowserExtension } from '@firebase/util';
@@ -61,7 +51,7 @@ export class AnalyticsService implements Analytics, _FirebaseService {
  * all or some of these, depending on the call's `send_to` param and the status
  * of the dynamic config fetches (see below).
  */
-let initializationPromisesMap: {
+export let initializationPromisesMap: {
   [appId: string]: Promise<string>; // Promise contains measurement ID string.
 } = {};
 
@@ -102,7 +92,7 @@ let gtagCoreFunction: Gtag;
  * Wrapper around gtag function that ensures FID is sent with all
  * relevant event and config calls.
  */
-let wrappedGtagFunction: Gtag;
+export let wrappedGtagFunction: Gtag;
 
 /**
  * Flag to ensure page initialization steps (creation or wrapping of
@@ -255,112 +245,4 @@ export function factory(
   const analyticsInstance: AnalyticsService = new AnalyticsService(app);
 
   return analyticsInstance;
-}
-
-/**
- * Sends analytics event with given `eventParams`. This method
- * automatically associates this logged event with this Firebase web
- * app instance on this device.
- * List of official event parameters can be found in the gtag.js
- * reference documentation:
- * {@link https://developers.google.com/gtagjs/reference/event
- * | the gtag.js reference documentation}.
- *
- * @public
- *
- * @param analyticsInstance - Firebase Analytics instance.
- * @param eventName - Event name to log. Can be a standard analytics event or any custom string.
- */
-export function logEvent(
-  analyticsInstance: Analytics,
-  eventName: string,
-  eventParams?: EventParams,
-  options?: AnalyticsCallOptions
-): void {
-  internalLogEvent(
-    wrappedGtagFunction,
-    initializationPromisesMap[analyticsInstance.app.options.appId!],
-    eventName,
-    eventParams,
-    options
-  ).catch(e => logger.error(e));
-}
-
-/**
- * Use gtag 'config' command to set 'screen_name'.
- *
- * @public
- *
- * @param analyticsInstance - Firebase Analytics instance.
- * @param screenName - Screen name to set.
- */
-export function setCurrentScreen(
-  analyticsInstance: Analytics,
-  screenName: string,
-  options?: AnalyticsCallOptions
-): void {
-  internalSetCurrentScreen(
-    wrappedGtagFunction,
-    initializationPromisesMap[analyticsInstance.app.options.appId!],
-    screenName,
-    options
-  ).catch(e => logger.error(e));
-}
-
-/**
- * Use gtag 'config' command to set 'user_id'.
- *
- * @public
- *
- * @param analyticsInstance - Firebase Analytics instance.
- * @param id - User ID to set.
- */
-export function setUserId(
-  analyticsInstance: Analytics,
-  id: string,
-  options?: AnalyticsCallOptions
-): void {
-  internalSetUserId(
-    wrappedGtagFunction,
-    initializationPromisesMap[analyticsInstance.app.options.appId!],
-    id,
-    options
-  ).catch(e => logger.error(e));
-}
-
-/**
- * Use gtag 'config' command to set all params specified.
- *
- * @public
- */
-export function setUserProperties(
-  analyticsInstance: Analytics,
-  properties: CustomParams,
-  options?: AnalyticsCallOptions
-): void {
-  internalSetUserProperties(
-    wrappedGtagFunction,
-    initializationPromisesMap[analyticsInstance.app.options.appId!],
-    properties,
-    options
-  ).catch(e => logger.error(e));
-}
-
-/**
- * Sets whether analytics collection is enabled for this app on this device.
- * window['ga-disable-analyticsId'] = true;
- *
- * @public
- *
- * @param analyticsInstance - Firebase Analytics instance.
- * @param enabled - If true, enables collection, if false, disables it.
- */
-export function setAnalyticsCollectionEnabled(
-  analyticsInstance: Analytics,
-  enabled: boolean
-): void {
-  internalSetAnalyticsCollectionEnabled(
-    initializationPromisesMap[analyticsInstance.app.options.appId!],
-    enabled
-  ).catch(e => logger.error(e));
 }
